% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{eRah_DB-class}
\alias{eRah_DB-class}
\title{Class \code{"eRah_DB"}}
\description{
The eRah_DB class contains the slots for storing and accessing a MS library.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the stored library}

\item{\code{version}}{The version of the stored library (and which is the database identifier, should be unique and used to check if is the database used in other experiments)}

\item{\code{info}}{Character vector containing complementary information about the library.}

\item{\code{database}}{A list of S3 objects, which each object contains the information on a different compound.}
}}

\author{
Xavier Domingo-Almenara.
}
