% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_methods.R
\name{expClasses}
\alias{expClasses}
\title{Experiment classes}
\usage{
expClasses(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object containing the experiment.}
}
\description{
The classes of a given experiment.
}
\details{
Returns the classes details of the experiment.
}
\seealso{
metaData phenoData
}
