% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSexport.R
\name{export2CEF}
\alias{export2CEF}
\title{Export spectra to CEF}
\usage{
export2CEF(Experiment, export.id = NULL, 
id.database = mslib, store.path = getwd())
}
\arguments{
\item{Experiment}{A 'MetaboSet' S4 object containing the experiment.}

\item{export.id}{If NULL, all the spectra in the experiment will be exported. Otherwise, only the AlignID in export.id will be exported}

\item{id.database}{The mass-spectra library used in the experiment.}

\item{store.path}{The path where the converted files are to be exported.}
}
\description{
Export spectra to CEF format for comparison with the NIST library through MassHunter interface.
}
