% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erah.R
\name{identifyComp}
\alias{identifyComp}
\alias{identifyComp,MetaboSet-method}
\title{Identification of compounds}
\usage{
identifyComp(Experiment, id.database = mslib,mz.range = NULL, n.putative = 3)

\S4method{identifyComp}{MetaboSet}(Experiment, id.database = mslib, mz.range = NULL, n.putative = 3)
}
\arguments{
\item{Experiment}{A 'MetaboSet' S4 object containing the experiment data previously created by newExp, deconvolved by deconvolveComp and optionally aligned by alignComp.}

\item{id.database}{The mass-spectra library to be compared with the empirical spectra. By default, the MassBank-[2] - Mass Bank of North America (MoNa) database are employed.}

\item{mz.range}{The same as in alignComp. If specified already in alignComp, then there is no need to especify it again. If not, it has to be specified.}

\item{n.putative}{The number of hits (compound candidate names) to be returned for each spectrum found.}
}
\value{
The function returns an updated S4 'MetaboSet' class, where the GC-MS samples have been now aligned.
}
\description{
Identification of compounds. Each empirical spectrum is compared against a ms library.
}
\references{
[1] Xavier Domingo-Almenara, et al., eRah: A Computational Tool Integrating Spectral Deconvolution and Alignment with Quantification and Identification of Metabolites in GC-MS-Based Metabolomics. Analytical Chemistry (2016). DOI: 10.1021/acs.analchem.6b02927 

[2] MassBank: A public repository for sharing mass spectral data for life sciences, H. Horai, M. Arita, S. Kanaya, Y. Nihei, T. Ikeda, K. Suwa. Y. Ojima, K. Tanaka, S. Tanaka, K. Aoshima, Y. Oda, Y. Kakazu, M. Kusano, T. Tohge, F. Matsuda, Y. Sawada, M. Yokota Hirai, H. Nakanishi, K. Ikeda, N. Akimoto, T. Maoka, H. Takahashi, T. Ara, N. Sakurai, H. Suzuki, D. Shibata, S. Neumann, T. Iida, K. Tanaka, K. Funatsu, F. Matsuura, T. Soga, R. Taguchi, K. Saito and T. Nishioka, J. Mass Spectrom., 45 (2010) 703-714.
}
\seealso{
\code{\link{newExp}} \code{\link{alignComp}} \code{\link{setAlPar}} \code{\link{setDecPar}}
}
\author{
Xavier Domingo-Almenara. xavier.domingo@urv.cat
}
