% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erah.R
\name{metaData}
\alias{metaData}
\title{metaData}
\usage{
metaData(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object containing the experiment.}
}
\description{
Displays the Experiment metadata
}
\seealso{
\code{\link{phenoData}}
}
