% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erah.R
\name{phenoData}
\alias{phenoData}
\title{Show phenotype data}
\usage{
phenoData(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object ciontaining the experiment.}
}
\description{
Displays the Experiment phenotypic data (if included).
}
\seealso{
\code{\link{metaData}}
}
