% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{RawDataParameters-class}
\alias{RawDataParameters-class}
\title{@name RawDataParameters-class
 @docType class
 @aliases RawDataParameters-class
 @title Class \code{"RawDataParameteres"}
 @description The RawDataParameters class contains the slots for storing and accessing into a MS sample, and the essential parameters for performing its processing (deconvolution).
 @slot data The data matrix of the sample to be processed
 @slot min.mz The minimum adquired mz number
 @slot max.mz The maximum adquired mz number
 @slot start.time Starting time of adquisition
 @slot mz.resolution Mz resolution
 @slot scans.per.second Scans per second
 @slot avoid.processing.mz Which mz do not have to be processed
 @slot min.peak.width Minimum peak width (stored in scans)
 @slot min.peak.height Minimum peak height
 @slot noise.threshold The noise threshold
 @slot compression.coef Compression coefficient (parameter for Orthogonal Signal Deconvolution)
 @author Xavier Domingo-Almenara.}
\description{
@name RawDataParameters-class
 @docType class
 @aliases RawDataParameters-class
 @title Class \code{"RawDataParameteres"}
 @description The RawDataParameters class contains the slots for storing and accessing into a MS sample, and the essential parameters for performing its processing (deconvolution).
 @slot data The data matrix of the sample to be processed
 @slot min.mz The minimum adquired mz number
 @slot max.mz The maximum adquired mz number
 @slot start.time Starting time of adquisition
 @slot mz.resolution Mz resolution
 @slot scans.per.second Scans per second
 @slot avoid.processing.mz Which mz do not have to be processed
 @slot min.peak.width Minimum peak width (stored in scans)
 @slot min.peak.height Minimum peak height
 @slot noise.threshold The noise threshold
 @slot compression.coef Compression coefficient (parameter for Orthogonal Signal Deconvolution)
 @author Xavier Domingo-Almenara.
}
