% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/support_methods.R
\docType{methods}
\name{expClasses}
\alias{expClasses}
\alias{expClasses,MetaboSet-method}
\title{expClasses-method}
\usage{
expClasses(object)

\S4method{expClasses}{MetaboSet}(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object containing the experiment.}
}
\description{
The classes of a given experiment.
}
\details{
Returns the classes details of the experiment.
}
\seealso{
metaData phenoData
}
