\name{expClasses}
\alias{expClasses}
\title{Experiment classes}
\description{The classes of a given experiment.}
\usage{
expClasses(object)
}
\arguments{
  \item{object}{A 'MetaboSet' S4 object containing the experiment.}
}
\details{
Returns the classes details of the experiment.
}

\seealso{
\code{\link{metaData}}, \code{\link{phenoData}}

}
