% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/erah.R
\docType{methods}
\name{metaData}
\alias{metaData}
\alias{metaData,MetaboSet-method}
\title{metaData-method}
\usage{
metaData(object)

\S4method{metaData}{MetaboSet}(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object containing the experiment.}
}
\description{
Displays the Experiment metadata
}
\seealso{
\code{\link{phenoData}}
}
