% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mslib.R
\docType{data}
\name{mslib}
\alias{mslib}
\title{MassBank Spectral Library}
\format{
An object of class \code{eRah_DB} of length 1.
}
\usage{
data(mslib)
}
\description{
The default mass spectral library of eRah, which is the MassBank repository.
}
\details{
This is the eRah default MS library, and automatically loaded with the eRah package. It contains almost 500 MS spectra. For details, see reference below.
}
\references{
[1] MassBank: A public repository for sharing mass spectral data for life sciences, H. Horai, M. Arita, S. Kanaya, Y. Nihei, T. Ikeda, K. Suwa. Y. Ojima, K. Tanaka, S. Tanaka, K. Aoshima, Y. Oda, Y. Kakazu, M. Kusano, T. Tohge, F. Matsuda, Y. Sawada, M. Yokota Hirai, H. Nakanishi, K. Ikeda, N. Akimoto, T. Maoka, H. Takahashi, T. Ara, N. Sakurai, H. Suzuki, D. Shibata, S. Neumann, T. Iida, K. Tanaka, K. Funatsu, F. Matsuura, T. Soga, R. Taguchi, K. Saito and T. Nishioka, J. Mass Spectrom., 45, 703-714 (2010)
}
\seealso{
\code{\link{compInfo}}
}
\author{
The TOF-MS spectra were contributted by Kazusa DNA Research Institute, the Engineering Department of Osaka University and Plant Science Center of RIKEN.

MassBank (http://www.massbank.jp/)
}
\keyword{datasets}
