% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/erah.R
\docType{methods}
\name{phenoData}
\alias{phenoData}
\alias{phenoData,MetaboSet-method}
\title{phenoData-method}
\usage{
phenoData(object)

\S4method{phenoData}{MetaboSet}(object)
}
\arguments{
\item{object}{A 'MetaboSet' S4 object ciontaining the experiment.}
}
\description{
Displays the Experiment phenotypic data (if included).
}
\seealso{
\code{\link{metaData}}
}
