% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/support_methods.R
\docType{methods}
\name{sampleInfo}
\alias{sampleInfo}
\alias{sampleInfo,MetaboSet-method}
\title{Information of the samples}
\usage{
sampleInfo(Experiment, N.sample = 1)

sampleInfo(Experiment, N.sample = 1)
}
\arguments{
\item{Experiment}{A 'MetaboSet' S4 object containing the experiment.}

\item{N.sample}{Integer. The number of the sample to query.}
}
\description{
Returns basic information on the samples.
}
\details{
Returns details on a given sample of the experiment, such as name, start time, end time, minium and maximum adquired m/z and scans per second.
}
\seealso{
\code{\link{plotChr}}
}
